+++ clang-tools-extra/clang-tidy/misrac/MisracTidyModule.cpp
@@ -13,6 +13,7 @@
 #include "Rule10p3Check.h"
 #include "Rule10p4Check.h"
 #include "Rule10p5Check.h"
+#include "Rule10p6Check.h"
 
 namespace clang::tidy {
 namespace misrac {
@@ -30,6 +31,8 @@ public:
         "misrac-rule-10p4");
     CheckFactories.registerCheck<Rule10p5Check>(
         "misrac-rule-10p5");
+    CheckFactories.registerCheck<Rule10p6Check>(
+        "misrac-rule-10p6");
   }
 };